feat: refactor eth form
