+++ packages/schemas/src/form/newForm.ts
@@ -14,7 +14,7 @@
 
 import { isAddress } from "viem";
 import { z } from "zod";
-import { address, addressOrEns } from "../web3";
+import { address, addressOrEns } from "../eth";
 
 // -----------------------------------------------------------------------------
 // Schema
